Move discount into price section
